requirements.txt
astpretty==3.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Deprecated==1.2.13
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
requests==2.28.1
smmap==5.0.0
urllib3==1.26.9
wrapt==1.14.1
nltk
parameterized
comby
pandas
jedi
z3-solver
numpy
scipy